Use direct remote plugin config sync
