+++ codex-rs/app-server/src/codex_message_processor/plugins.rs
@@ -1,13 +1,10 @@
 use super::*;
-use crate::config_write_router::RemotePluginEnablementWriter;
 use crate::error_code::internal_error;
 use crate::error_code::invalid_request;
-use async_trait::async_trait;
 use codex_app_server_protocol::PluginInstallPolicy;
 
-#[async_trait]
-impl RemotePluginEnablementWriter for CodexMessageProcessor {
-    async fn set_remote_plugin_enabled(
+impl CodexMessageProcessor {
+    pub(crate) async fn sync_remote_plugin_enabled_config_write(
         &self,
         plugin_id: String,
         enabled: bool,
@@ -18,9 +15,7 @@ impl RemotePluginEnablementWriter for CodexMessageProcessor {
         {
             return Err(invalid_request("remote plugin enablement is not enabled"));
         }
-        if plugin_id.is_empty()
-            || !codex_core_plugins::remote::is_valid_remote_plugin_id(&plugin_id)
-        {
+        if plugin_id.is_empty() || !is_valid_remote_plugin_id(&plugin_id) {
             return Err(invalid_request(
                 "invalid remote plugin id: only ASCII letters, digits, `_`, `-`, and `~` are allowed",
             ));
@@ -49,9 +44,7 @@ impl RemotePluginEnablementWriter for CodexMessageProcessor {
         self.clear_plugin_related_caches();
         Ok(())
     }
-}
 
-impl CodexMessageProcessor {
     pub(super) async fn plugin_list(
         &self,
         request_id: ConnectionRequestId,
@@ -455,9 +448,7 @@ impl CodexMessageProcessor {
                 "remote plugin install is not enabled for marketplace {remote_marketplace_name}"
             )));
         }
-        if plugin_name.is_empty()
-            || !codex_core_plugins::remote::is_valid_remote_plugin_id(&plugin_name)
-        {
+        if plugin_name.is_empty() || !is_valid_remote_plugin_id(&plugin_name) {
             return Err(invalid_request(
                 "invalid remote plugin id: only ASCII letters, digits, `_`, `-`, and `~` are allowed",
             ));
@@ -626,13 +617,13 @@ impl CodexMessageProcessor {
     ) -> Result<PluginUninstallResponse, JSONRPCErrorError> {
         let PluginUninstallParams { plugin_id } = params;
         if codex_core::plugins::PluginId::parse(&plugin_id).is_err()
-            && !codex_core_plugins::remote::is_supported_remote_plugin_id(&plugin_id)
+            && !is_valid_remote_uninstall_plugin_id(&plugin_id)
         {
             return Err(invalid_request(
                 "invalid plugin id: expected a local plugin id in the form `plugin@marketplace` or a remote plugin id starting with `plugins~`, `app_`, `asdk_app_`, or `connector_`",
             ));
         }
-        if codex_core_plugins::remote::is_supported_remote_plugin_id(&plugin_id) {
+        if is_valid_remote_uninstall_plugin_id(&plugin_id) {
             return self.remote_plugin_uninstall_response(plugin_id).await;
         }
         let plugins_manager = self.thread_manager.plugins_manager();
@@ -715,9 +706,7 @@ impl CodexMessageProcessor {
         {
             return Err(invalid_request("remote plugin uninstall is not enabled"));
         }
-        if plugin_id.is_empty()
-            || !codex_core_plugins::remote::is_valid_remote_plugin_id(&plugin_id)
-        {
+        if plugin_id.is_empty() || !is_valid_remote_plugin_id(&plugin_id) {
             return Err(invalid_request(
                 "invalid remote plugin id: only ASCII letters, digits, `_`, `-`, and `~` are allowed",
             ));
@@ -739,6 +728,21 @@ impl CodexMessageProcessor {
         self.clear_plugin_related_caches();
         Ok(PluginUninstallResponse {})
     }
+}
+
+fn is_valid_remote_plugin_id(plugin_name: &str) -> bool {
+    plugin_name
+        .chars()
+        .all(|ch| ch.is_ascii_alphanumeric() || ch == '-' || ch == '_' || ch == '~')
+}
+
+fn is_valid_remote_uninstall_plugin_id(plugin_name: &str) -> bool {
+    !plugin_name.is_empty()
+        && is_valid_remote_plugin_id(plugin_name)
+        && (plugin_name.starts_with("plugins~")
+            || plugin_name.starts_with("app_")
+            || plugin_name.starts_with("asdk_app_")
+            || plugin_name.starts_with("connector_"))
 }
 
 fn remote_marketplace_to_info(marketplace: RemoteMarketplace) -> PluginMarketplaceEntry {
